+++ statsd/client/__init__.py
@@ -2,3 +2,4 @@ from __future__ import absolute_import, division, unicode_literals
 
 from .stream import TCPStatsClient, UnixSocketStatsClient  # noqa
 from .udp import StatsClient  # noqa
+from .consistent_hashing import ConsistentHashingStatsClient  # noqa
